Crate cumulus_client_parachain_inherent

Client side code for generating the parachain inherent.

Structs

MockValidationDataInherentDataProvider

Inherent data provider that supplies mocked validation data.

MockXcmConfig

Parameters for how the Mock inherent data provider should inject XCM messages. In addition to the messages themselves, some information about the parachain’s configuration is also required so that the MQC heads can be read out of the parachain’s storage, and the corresponding relay data mocked.

ParachainInherentData

The inherent data that is passed by the collator to the parachain runtime.

ParachainInherentDataProvider

Constants

INHERENT_IDENTIFIER

The identifier for the parachain inherent.